update finall All testing

+++ src/main/java/testing/calculator/Operation.java
@@ -13,7 +13,7 @@ public class Operation {
         return number1 * number2;
     }
 
-    public double  divide(double number1, double number2) {
+    public int  divide(int number1, int number2) {
         return number1 / number2;
     }
 }
